프레임 수집 로직 개선 및 데이터 타입 불일치 문제 수정
- 컬렉션 종료 조건을 `시간 기반`으로 변경, 안전 정지 로직 추가 (MAX_FRAMES_TO_COLLECT 도입)
- main.py: 프레임 수집 기간 상수 추가 (COLLECTION_DURATION_SECONDS)
- predictor.py: 패딩 데이터 타입 불일치 문제 수정 (float32 명시)
- WebSocket 구현:
  - 메시지 수신 시 컬렉션 타이머 시작, 데이터 추가 시 조건 검증 강화
  - 수집 완료 시 정확한 inference 실행 로직 구현 및 메시지 전송
- /config 엔드포인트 확장: 추가 설정값 반환 (COLLECTION_DURATION_SECONDS, MAX_FRAMES_TO_COLLECT)
- 안정성 및 메모리 관리를 위한 Collector 클래스 리팩터링

+++ processing/predictor.py
@@ -188,9 +188,10 @@ class Predictor:
         max_frames = 300 # 학습 시 사용한 max_frames와 동일해야 함
         if num_frames < max_frames:
             pad_size = max_frames - num_frames
+            # 패딩 배열의 데이터 타입도 float32로 명시하여 타입 불일치 문제를 방지합니다.
             landmarks_np = np.vstack([
                 landmarks_np,
-                np.zeros((pad_size, landmarks_np.shape[1]))
+                np.zeros((pad_size, landmarks_np.shape[1]), dtype=np.float32)
             ])
         elif num_frames > max_frames:
             landmarks_np = landmarks_np[:max_frames]
